Crate disassemble

Disassemble

This crate provides basic functionality for working with disassembled code.

The actual disassembly with the implementation of Instruction and other elements of the system will be provided by other crates that integrate with other systems, such as the Capstone Engine.

It is possible (likely?) that some functionality from within this crate may move in the future to a separate crate for broader re-use. This might impact Symbol among other things.

Structs

Address
The location of something in an address space.

BasicBlock
A basic block is a sequence of instructions with no inward-bound branches except to the entry point and no outward-bound branches except at the exit.

Function
A function within a program.

Symbol
A symbol within an executable or library. This is a named address.

Traits

Instruction
An assembly instruction, bytecode operation, VM operation, etc.