Struct disassemble::BasicBlockEdge

pub struct BasicBlockEdge<'f> {
    pub edge_type: EdgeType,
    pub direction: EdgeDirection,
    pub other_bb: &'f BasicBlock<'f>,
}

Information about an edge between 2 basic blocks.

This represents a branch, jump or other form of control flow transfer within the control flow graph.

Fields

edge_type

Is this edge taken conditionally or unconditionally?

direction

Is this an inbound or outbound edge?

other_bb

What is the other BasicBlock involved in this edge?

If the direction is EdgeDirection::In, then this other_bb will be the source bb and this bb will be the destination.

If the direction is EdgeDirection::Out, then this bb will be the source and other_bb will be the destination.