Struct disassemble::BasicBlockEdge

pub struct BasicBlockEdge {
    pub edge_type: EdgeType,
}

Information about an edge between 2 basic blocks.

This represents a branch, jump or other form of control flow transfer within the control flow graph.

Fields

edge_type

Is this edge taken conditionally or unconditionally?